+++ ManaKit/Classes/DeckTableViewCell.swift
@@ -58,21 +58,19 @@ public class DeckTableViewCell: UITableViewCell {
     
         // thumbnail image
         if let card = deck.heroCard {
-            if let croppedImage = ManaKit.sharedInstance.cardImage(card,
-                                                                   imageType: .artCrop,
-                                                                   faceOrder: faceOrder,
-                                                                   roundCornered: false) {
+            if let croppedImage = card.image(type: .artCrop,
+                                             faceOrder: faceOrder,
+                                             roundCornered: false) {
                 thumbnailImage.image = croppedImage
             } else {
                 thumbnailImage.image = ManaKit.sharedInstance.imageFromFramework(imageName: .cardBackCropped)
                 
                 firstly {
-                    ManaKit.sharedInstance.downloadImage(ofCard: card, imageType: .artCrop, faceOrder: faceOrder)
+                    ManaKit.sharedInstance.downloadImage(ofCard: card, type: .artCrop, faceOrder: faceOrder)
                 }.done {
-                    guard let image = ManaKit.sharedInstance.cardImage(card,
-                                                                       imageType: .artCrop,
-                                                                       faceOrder: self.faceOrder,
-                                                                       roundCornered: false) else {
+                    guard let image = card.image(type: .artCrop,
+                                                 faceOrder: self.faceOrder,
+                                                 roundCornered: false) else {
                         return
                     }